build: add "check" step
This adds the "check" step into the library, which builds, but doesn't
run, unit tests.

This allows zls to do further semantic checking on the code by enabling
build-on-save.

Note that since we are still on 0.13.0, users of zls still need to
explicitly enable the "check" step in their editor to use this, but with
0.14.0, this requirement goes away, so relatively soon this should
"just work".

This has already been great for visual feedback for me, so likely expect
a lot more unit tests in the future as it means I won't have to run "zig
build test" on the CLI ad nauseam anymore.