gradient: add radial gradients
This adds support for 2-point radial gradients to the library.

Radial gradients are defined by two circles that make a cone if you were
to look at it from 3D space; the observer looks at the cone top-down
with the inner circle at the top. This is a well-defined type of radial
gradient style, and reduces to a simple radial gradient if the inner
radius is chosen to be zero with the same co-ordinates for both the
inner and outer circles.

Our algorithm has been imported from Pixman which has a good proof of
the algorithm and easy to follow code. Note that there may be some
differences between what we render versus Pixman on some edge cases due
to Pixman's use of fixed point, but some testing to check has shown that
we are either close or identical in the nominal cases at least.

As part of this work to make sure we're in line with Pixman, linear
gradients have had their co-ordinates adjusted so that they evaluate the
pixel from the center (x/y += 0.5); this was previously not done as it
was causing results that I was not expecting during unit tests, but I
believe accuracy with an existing library is preferable in this
situation (and helps if we want to test against Cairo).

Additionally, this commit introduces the convention of returning
negative values from "getOffset" calls to denote invalid t values;
certain radial cases return this in addition to now degenerate linear
gradients where division by zero might occur. When sent to a stop
search, these invalid cases return transparent black.