gradient: add conic gradients
This adds conic gradients (the last gradient type for now) to the set of
supported gradients.

Conic (or "sweep") gradients are probably the simplest algorithm of all
the gradients so far (and probably out all the ones we will implement)
since they just work off a starting point and an offset angle. For any
given point, the angle is calculated from the slope of point minus
center, offset by the supplied angle, and then lerped according to its
angle within the circle, clockwise.

As part of this, there's been a few fixes added:

* HSL colors now allow 360 to be specified as an appropriate hue value
  to allow defining correct stops along the long path from red -> red
  (ie, 0 to 360 going across the full circle).
* Identical offset stop behavior has been changed for searching, I
  originally though we needed to rewind state to get the right stops
  returned for hard stops, but it turns out this was incorrect and was
  actually giving incorrect behavior.
* Added some notes about smoothing for hard stops, since supersampling
  does not apply to pixel sources. This sees to actually be an issue for
  several graphics libraries, including Skia (issue put in one of the
  acceptance tests), that manifests itself within CSS. The workaround in
  CSS is usually to just add a small offset on one side of the hard stop
  to add a small amount of interpolation. Obviously, with access to
  the lower-level primitives within the library itself other options are
  available, but we assert that the offset case works for us too.