gradient: add license statement to radial gradient algo comments
Different parts of Pixman (actually, lots of Pixman :P) have
terms/language that more resemble the Prior BSD License versus the MIT
license as seen in the COPYING file. While the notice is a bit meaty I
don't necessarily see the issue with adding the full statement to the
comments breaking down the radial gradient algorithm, so I've done that
now.