gradient: add transformations
This sets up the ability to apply transformations on gradients.

This also sets up the codification of the flow of co-ordinate space from
pattern -> user -> device, similar to Cairo; meaning that the
expectation is that any transformation that you apply to a gradient is
actually saved as the inverse, and to get the desired effect you need to
draw with generally the same transformations in place.