Dither: new low-level pattern type
This adds support for the Dither pattern, a new pattern type that is
only available through raw composition, context, and the painter
(context bits will be added in another commit).

The current plan is to just support ordered dithering for the time
being, although in theory the dither pattern should be flexible enough
to add other, non-ordered types eventually.

Additionally, only 8x8 Bayer is currently in. This is what we will
probably be launching with in 0.6.0, however there is also plans to add
64x64 blue noise/void and cluster, which should be able to just be
dropped in when we have a pre-generated matrix.