dither: add blue noise matrix
Famous last words indeed from a couple of days ago - I've decided to add
the blue noise matrix after all.

Pixman has done the heavy lifting for us, so we don't need to generate
our own and can just use theirs. Source mentioned, terms in the LICENSE
file (neither the header file nor the generator file actually had the
prior BSD license header, yay!).

We also roll in the gather function from
https://github.com/ziglang/zig/issues/12815#issuecomment-1243043038.
Once Zig actually has first-class gather, we can switch it to that.
There might be other place where we can implement this too (we kind of
do some gather stuff in the compositor, but I'm not in a rush to fix
that at this point as we already do a for (0..vector_length) in a lot of
those with the appropriate indexing, so it might be compiler equivalent,
we'd have to see.