compositor: add support for different compositor operators
This adds support for several different compositor operators; 19 new
operators, for a current total of 21!

These come from different sources, namely porter-duff and
PDF-specification blend modes. We don't necessarily discern between the
two; but they are aptly named, so those familiar with the different
kinds of operators and blend modes should understand exactly what they
mean.

As part of this, when filling and stroking, the operator to use can now
be chosen with "Context.setOperator" and a respective configuration
field in "painter.fill" and "painter.stroke". This affects the operator
used during the final composition of the mask on the surface.

Consistent to their documented behavior in Cairo, ".src_in", ".dst_in",
".src_out", and ".dst_atop" are also unbounded, meaning that when they
are used in fill or stroke operations, they remove anything outside of
their affected area; this is done via a clear operation in direct paint
and just by increasing the mask surface to the whole of the destination
surface in the event of AA paint.

Finally, we have renamed the original operators, ".in" is now ".dst_in",
and ".over" is now ".src_over".

There are a number of operators (color dodge, color burn, soft light,
and non-separable HSL blend modes) that are not a part of this initial
batch, as they require complex calculations and a higher-degree of
precision than can be achieved in the integer pipeline. These will be a
part of future higher-precision floating-point pipeline.