compositor: add floating-point precision, add remaining operators
This adds floating-point precision support to the compositor (we just
re-use the color.LinearRGB functionality for this), and support for the
remaining operators that required floating-point support.

The remaining operators were: .color_dodge, .color_burn, .soft_light,
.hue, .saturation, .color, and .luminosity. This brings the total
supported set of operators to 28, pretty much in line with what you'd
expect in a standard graphics library (all standard porter-duff and
PDF blend modes are now supported).

Precision will be automatically switched to floating-point if an
operator that requires it is detected in fill and stroke, and also with
surface-level compositor operations. Using the stride-level compositor
directly, however, will not upgrade you. Any operator that requires
floating-point support in this latter situation will just clear the
affected area.