compositor: vectorize partial end-of-stride pixels
As it stood, the compositor was relying on scalar API for the last parts
of a stride (read: line) that were not big enough for a vector. This
actually leaves a decent amount of performance on the table, especially
on systems where the vector size can grow to super large amounts (think
AVX2, AVX512, etc).

This commit allows partial vectors to be loaded on all respective vector
load/store functions using a comptime flag and a length, the latter of
which is only checked if the flag is true. Out of all the things that I
tried (this, code duplication, and inlining), this seems to be the best
for binary size and maintainability. The comptime flag should ensure
that length constraints are only checked if necessary, otherwise,
they should be optimized out.

Note that this now leaves the scalar API unused the compositor. I'm
going to hold off on reviewing the possibility of taking it out until
after 0.6.0. It's also useful for testing, and I'm wondering if it's
possibly going to be useful elsewhere, or useful for external consumers
still (their existence allows consumers to use several primitives
directly, which is something we might want to continue to support).