Support PNG export gamma, add sRGB interpolation space
This commit adds the ability to the select the RGB profile to export as
in the toy PNG exporter, by adding an options parameter to the export
function.

The parameter is optional and the default is null, which falls back to
the default behavior of not re-encoding image data or setting the gAMA
header to the correct gamma setting.

When set, the gAMA header is set accordingly to the gamma transfer
function, in addition to the data being gamma-encoded according to the
space selected.

Also, mainly for compatibility's sake, the .srgb interpolation color
space has been added. This allows you to, say, set all of your colors to
sRGB, and also set the interpolation color space as sRGB, and then
expect pretty much the same result when exporting in the sRGB space when
compared to doing everything in linear. Note that this may not be what
you want, and to get smoother transitions, you probably still want to
interpolate in linear RGB even if you are working with sRGB.

The acceptance tests specifically demonstrate this (depending on the
image viewer you use; some viewers will try to automatically adjust and
both new acceptance tests my look the same as a result).

Finally, note that our transfer function is currently the fast transfer
function (inverse of pow(2.2)), and may not exactly line up with true
sRGB, especially on very dark colors. We may add options to accommodate
this later.

It should also be noted that only RGB and RGBA pixel types currently
support changing the gamma when exporting.