Move vectorized color and dither API to internal
The vectorized API in the color (and ultimately stuff that was needed in
Dither) was never meant to be public but a ton of it was needed for the
compositor. I wasn't the biggest fan of having it like that, so I've
moved it to an internal package now so it's pretty clear it's not
supposed to be used externally.