Introduce text support
This commit is the initial support for text in z2d. It introduces 2 new
components and allows for basic Latin text rendering, including
diacritics (via composite glyphs, which some fonts use for such
situations) and kerning.

* "text" is the unmanaged text rendering package, which allows one to
  draw text in a fashion similar to how you would use "painter" to fill or
  stroke generally. Rendering text is technically a fill operation, so
  fill options are passed down via "text.show" as well.

* "Font" controls the loading and encapsulation of TrueType/OpenType
  fonts, glyphs, and outlines. The latter two are located in an internal
  file, as they are not necessarily designed to be interfaced with
  externally. Fonts need to be, however, so that a font can be loaded
  for use with the unmanaged interface.

This functionality is shipped to Context via the "setFontToFile",
"setFontToBuffer", "setFontSize", and "showText" methods.

Note that while we use a file-style I/O in our implementation of fonts.
This allows us to not only lean on appropriate parts of stdlib like
readInt et al. to easily interpret memory, but provides us with a safety
mechanism against reading beyond the end of a font file or similar other
issues. Additionally, we verify table checksums upon loading a font
file.

Considering the above, note that we still load the file into memory in
its entirety upon initialization, using FixedBufferStream to wrap the
memory. This can be externally managed using the appropriate loadBuffer
and setFontToBuffer functions, which allows for the possibility of
advanced scenarios such as mmap.

Certain things are not supported yet - namely, non-Latin,
non-left-to-right text is not support at this moment, nor are font
collections. These are things that are not off the roadmap and will be
done eventually!