surface: widen some dimensions for calculations/compositing
This ensures that some dimensions that were not widened during
compositing and surface allocation (ImageSurface allocation
particularly) are now properly done.

Note that this was (unintentionally) limiting memory that could be
allocated for ImageSurfaces to ~2GB, and now this limitation is gone, so
folks should understand that and be careful. :)

Note that in applications where this is likely to come up (rendering
large surface areas with AA), there's likely improvements that can be
made here, although they are outside the scope of this small fix.