Slope: add epsilon tolerance for compare
This adds some tolerance for slopes where the dy or dx only differ by
the f64 epsilon or less. If that is the case, we snap the second dx, dy,
or both, to the respective component of the first slope so they compare
equal.

This prevents some cases that ultimately calculate to degenerate ones
from producing NaNs during stroke plotting.