Polygon: assert on finite points on plot and plotReverse
Every now and then we catch a degenerate case that generates NaN/Inf
point values, which are currently a bit hard to track down because
they're usually caught during rasterization, when the floating-point
value is cast to integer. By this point, the polygon is fully "built"
and the exact point where it happens is lost.

Hence, this is being added just as a safety/debugging catch-all to make
sure no NaNs/Infs make it into a polygon at plot-time, which should make
these cases easier to track.