painter: validate that polygon lies within surface before compositing
This is the start of some validation work that I am undertaking to make
sure edge cases are better validated against.

The first one is a fairly simple one in that we now will check to make
sure that a polygon actually lies within the surface in some respect (in
full, or overlapping) before drawing. Not only does this provide better
protection against these kinds of cases, it skips unnecessary work in
some cases, like in the event the polygon is adequately y-aligned but
still outside on the x-axis.