surface: disallow zero-size dimensions
We currently only enforce a limitation on size dimensions when
specifying surface.init and others to ensure non-negative values do not
get provided. However, it's still possible to provide a zero to height
or width, and get a zero-length surface, with some odd characteristics
(namely, one of height or width will be non-zero).

To ensure correctness here, we're now enforcing that dimensions need to
be positive, non-zero numbers.