stroke_plotter: handle special case closes where current == initial
This handles the special case where the current point in a stroke is the
same as the initial point. This case is most likely to come up if the
a line_to is processed to initial point before the close. In this case,
we just treat the last join (before the closing one) as being already
done, and just make sure we take into account the shifting of the points
when we do the closing join.