surface: ensure downsampling aborts when dimensions too small
I noticed this when reviewing the downsampling functionality recently.
Given that we have now made a note that zero-length buffers are no
longer a thing, we should ensure that downsampling cannot downsample a
buffer into being empty, so this adds that now.

Also added some tests and documented functionality with regards to what
happens when a buffer with uneven dimensions is downsampled.