Make vectors static and configurable
This changes the vector size to be static and build-configurable, with a
default of 16 (so targeting 256-bit registers, e.g., AVX2).

The value can now be configured by adding a .vector_length value to the
dependency args when loading via build.zig.

Configuration of this value was something that I always wanted to do.
Moving away from simd.suggestVectorLength was also something that I've
wanted to do for a bit too, as I realize that a number of projects may
not specify the necessary target features to be able to take advantage
of larger vector sizes, and the build process would be likely to select
a smaller vector size than needed. This fixes both of the issues as the
default vector size should now be reasonable for the majority of
consumer CPUs out at the moment, with operations lowering to multiple
ones when working with platforms that have smaller vector sizes.
Additionally, having this option build-configurable allows manual
tweaking for specific environments and targets.