build: add optimize option to module instance
This is just for consistency's sake; we added ".target" since we are now
using the module instance (instead of a direct root source) for our unit
tests (we were already doing this for spec tests) and it was complaining
about a lack of target. Additionally, it seems like while currently,
optimization mode is completely inherited from the root module of any
particular build (e.g., via "addExecutable" or "addTest"), more granular
control in the future will be implemented.