Polygon: re-introduce xEdgesForY i30 clamping
When calculating, the x-edges returned by the (previously named)
edgesForY method would be clamped to i30 (or should have been, after
reviewing the code I'm not too sure if this would have actually worked).
This is because we store the edges in a small stack-based buffer (512
bytes) that also includes the direction, and we want to make sure that
there is enough room for 16 items here so these values are packed.
However, the clamping was lost in #128.

This re-introduces it and ensures that the clamping happens correctly
before conversion, which was the thing that was not happening even
before #128. A unit test is also added that will trigger safety-checked
UB if the clamping is not happening.