painter: ditch FBA for x-edge allocation
This switches the FBA that we were using for x-edge allocation during
the scanline fill for a simple arena wrapping the allocator that is
passed in at the higher level.

This is being done so that we have more freedom in determining the data
that gets passed back during x-edge calculation. Currently, we make
efforts to try and fit the entirety of a single edge into a u32
(including truncating the actual x-coordinate to an i30, which I want to
get rid of). Future AA algorithms may require that we keep a reference
back to the actual edge data (so that we can further calculate
sub-scanlines during rasterization) - this will increase the amount of
data we need to keep to at least 96 bits, and 128 if we don't pack it.
This either requires more data to go on the stack, or less edges be
stored.

Benchmarks have shown that the cost of removing the FBA is fairly
minimal. Additionally, the use of an arena allows us to re-use the same
memory every scanline, keeping the actual allocations that happen down
to a minimum.