Polygon: remove i30 restriction on x-edges, remove x-edge iterator
This removes the i30 restriction when selecting x-edges during the
scanline rendering process of fill and stroke.

Additionally, we remove the x-edge iterator, and instead move non-zero
rule filtering to the actual x-edge discovery process instead.

Finally, only the edges are returned now (no unneeded metadata like
direction).

To accommodate this without affecting performance, we're internally
using a MultiArrayList, which allows us to maintain the association
between edge x-coordinate and the direction of the edge that
x-coordinate came from. This allows us to still sort according to the
x-coordinate and then filter in the correct order. We drop the direction
field after the filter is complete.