painter: implement multi-sample anti-aliasing
This adds multi-sample anti-aliasing (MSAA) as a new anti-aliasing
method.

Our algorithm employs a scanline-based sparse coverage buffer, similar
to what is in employ by the tiny-skia project, and assuming Skia, before
they switched to analytical anti-aliasing, although the sparse buffer
could still exist, I haven't checked too much. ;p

The employ of the sparse buffer and coverage method, however, fits in
very nicely and iteratively with our existing supersample anti-aliasing
(SSAA), in that we can still supersample co-ordinates at 4x and
scanline-raterize in the same fashion, but rather than employing a
4x-size mask sized for the entire draw area, we employ a 1x buffer sized
to a single scanline. This provides amazing and immediate memory
savings - for draw areas smaller 256 pixels wide, we only use a maximum
of 510 bytes of memory (2 []u8 buffers, one for coverage, one for
lengths, with the latter conditionally sized to only be as wide as we
would need in co-ordinate space). Compare with 4x SSAA - 255x255x4 =
260100 bytes, saving nearly 260K!

Note that this savings favors vertically-skewed drawing of course -
shapes that are longer on the x-axis will see diminished returns in
memory, but should still see savings in speed due to the sparse nature
of the buffer (and for worst-case scenario for most reasonably-sized
even single-pixel high lines, will only use 75% of the mask storage SSAA
would).

This also has the effect of rendering exactly the same as SSAA in nearly
every case; currently, only a few sub-u8-sized depth images render very
slightly differently (even in these cases, the difference is almost
unnoticeable).

Speed is also much improved; results vary from scenario to scenario, but
we are looking at as much as over 2x increase in speed in a number of
tests using SmpAllocator. Note that this time gap does close using
c_allocator, which could be a testament to the role that memory pressure
is playing here. However, even under this scenario, speed increases can
still be between 1.1x - 2x.

As such, given that there are no drawbacks to using this method that
have surfaced, and only gains, it will be the default anti-aliasing
method for the next release (0.7.0), although we will be switching the
.default alias in another commit.

Included with this commit is a new benchmark framework that works off
the existing acceptance tests. It's contained in a subproject in spec/
to avoid polluting the main package and to ensure we can easily include
zbench. I eventually want to re-add profiling using this same framework.