Path: lower co-ordinate limit to i24
This commit lowers our path co-ordinate limit to i24 - meaning that any
co-ordinate set using moveTo, lineTo, etc., will now be limited to
+/-8388608.

This is something I've mulled on and off for probably as long as z2d has
existed, mainly as a measure of ensuring safety in the event of changes
to accommodate things like fixed-point, which has never materialized as
being necessary or useful, although this may change and has not been
completely exhausted as a possibility.

Before this, the limit was i32, which was mainly serving as a upper
limit of what could ultimately be returned in the event of drawing when
no anti-aliasing was being done. However, anti-aliasing *does*
effectively reduce the reasonable range to +/-536870912, and the advent
of MSAA raises the possibility that we could go even lower, possibly
from 4x to 8x or even 16x, so possibly as low as +/-134217728. However,
even at these limits I'd be very worried for things like performance or
other constraints (even though we do try as hard as possible to prevent
overflows and what not during rasterization calculations).

Rather than speculate over what the future holds or what the real-world
scenarios look like, I've decided to take the safe route and clamp to
i24. This provides future-proofing in the event that we use fixed point
(even if it's just in a limited fashion, like during scanline
rasterization, as opposed to tessellation, where if we don't necessarily
benefit from integer calculations, we still benefit from the savings on
the values we need to hold in memory), while also setting a reasonable,
albeit still very extremely liberal, limit on the co-ordinate space,
that should accommodate any foreseeable need of the library.

For some context and prior art:

* Cairo effectively clamps to i24 (they make heavy use of 24.8 fixed
  point). This can be confirmed by looking at the value of
  CAIRO_FIXED_MAX_DOUBLE.
* I've seen some mention of Skia clamping to i16, but I have not 100%
  confirmed.
* Pixman is in a similar scenario where I've seen mentions of clamping
  to i16, but more interrogation would be needed to 100% confirm.