Update to latest Zig, post major I/O changes
This updates Zig to the latest version (0.15.0-dev.1509+b9a6dae2a).

This is also the first update after the major I/O changes that have been
going on the for the last couple of weeks. This mainly affects us in our
PNG export code, namely due to the fact that the zlib compression
library has been removed due to what I'm assuming is an inherent
inflexibility that has made it infeasible to refactor.

In the meantime, I've vendored some relevant pieces that we need to do
support the current PNG export functionality. A large amount of the
functionality has been pruned, so mostly only what we need currently
should remain. The compression part should of PNG exporting should be
stable, so I don't expect needing to add anything else, and hopefully we
can remove this entirely in due time in favor of a new stdlib
implementation.