fill_plotter: no-op degenerate close_path cases
This removes the InvalidState error when there are less than 3 points in
the buffer when close_path is called in the fill plotter. This ensures
that duplicate closePath calls succeed, and also that degenerate moveTo
-> lineTo cases also work. In the latter case, an edge will be
generated, which will likely draw incorrectly, but the rasterizer will
still handle this gracefully as odd edges at the end of a scanline are
discarded.