sparse_coverage: don't maxInt-limit len on put
This changes put on the sparse coverage buffer so that the len value
will no longer be explicitly limited by the max call. I've added an
assert too just to double it up, just to make sure you're not putting a
length past capacity.

While neither an explicit limit, nor implicit truncation (via intCast)
is what we want in this situation (and would lead to invalid lengths in
the buffer), the latter will at least trigger a safety check, and after
giving it some thought, I think this is more the correct way to go.

Note that this doesn't change anything regarding the lack of safety
guarantee of the sparse coverage buffer and we still expect safety
checks to happen farther up the stack. In fact, this very situation
should be blocked two-fold in MSAA, one in the sub-scanline loop via
correct limits on spans, and the other in walking the length of the
buffer as we limit the walk to the capacity of the buffer at most.