Remove new file I/O buffers from a couple of places
This removes the 16K buffers added in a couple of places to (and an
8-byte one in one place too).

These were added to support the new I/O framework, before I knew that a
zero-length buffer (i.e., no buffer) was valid, and upon further thought
I think this is the more correct position for what we're doing in these
situations - file I/O where hopefully there would be no need for our own
buffering and we can just rely on the filesystem for said facilities. If
this is wrong we can add them back, but otherwise I'd rather not have
16K on the stack laying around in a couple of places, and for lack of a
better idea on a good size for these scenarios, I think no buffer and
just streaming is okay.