Text rendering fixes and new logo
This adds a new logo - and a few fixes for text rendering that I found
in making said logo!

The logo was done entirely in z2d and includes a working example of our
new text rendering functionality. The logo is a variation of the actual
Zig logomark, with the typography used for the additions to the main
logo, and following subtitle, being Montserrat, the font that the
textual Zig logos derive from. This gives us a minimal and derivative
logo that should (hopefully) evoke a sense of familiarity from those
that work with Zig, which I think is suitable for our library.

As mentioned, during the process of making this logo, and trial and
error testing various fonts and what not, I encountered some bugs which
are fixed by this commit as well (and some of the errors were discovered
through using Montserrat, so the new logo test also helps tests against
that):

* RLS overflow issues when loading short glyph indexes - the short
  indexes are packed and actually represent a possible u32 index / 2,
  but we were not widening before doing the multiplication.
* Some offset data during outline processing for glyphs was being
  incorrectly read into i8 fields using readByte() instead of
  readByteSigned(), causing overflows.
* Switching fonts in the context was incorrectly trying to free
  buffer-only fonts (as if they had been read in off of disk).

These issues are now fixed.

Additionally, we are now exporting Context.deinitFont().