Make deinit calls more strict/correct
The correct deinit convention should be: release resources, invalidate
memory. We are currently not following the latter part of that
convention. This fixes that, and sets the expectation for future deinit
calls. :)

Note that there's Context.deinitFont that does a "pseudo-invalidation" -
I've considered renaming this function (briefly, while I'm typing this
;p ), and I think it's fine as-is. It *can* be considered an
invalidation as it sets the in-use font to .none, which makes using text
functions invalid until another font is set. That's following the spirit
of the convention, as far as I'm concerned.