Simplify remove buffer-less R/W in favor of direct file I/O
This removes the places we had buffer-less writers (the ones we switched
in 54fb7e3191f51a5319993629077248f37c14aeab), switching them out for
direct file I/O in these situations.

No buffer ultimately meant we were doing this anyway - read directly
calls to this, and write will do this with extra steps on a zero-length
buffer. As such, this ultimately will cause less confusion, and the
following is much more clear:

* Font file reading will be done directly into the pre-allocated buffer
  that serves as the underlying data that exists for the lifetime of a
  Font loaded by Font.loadFile.
* For PNG exporting, there's two places of concern - the PNG header, and
  chunk encoding. Header encoding is trivial. Chunk encoding entails a
  few bytes of metadata (12) surrounding the actual chunk payload. The
  main concern here would be IDAT chunks, but we are pretty well
  serviced here by our existing 16KiB zlib buffer, meaning that a
  moderately small amount of IDAT chunks get written out anyway
  (currently 7 for our largest file in spec/, which comes in at ~103KB).

I think the initial confusion, and why we didn't just switch to this
pattern right away after the large change to std.Io, came from trying to
figure out where some convenience functions like readAllAlloc and
writeInt went; this is solved pretty easy by allocating ahead of time
(in the case of readAllAlloc) or local implementations (in the case of
writeInt).

I find where this is at now much more clear. For any future thing that I
may find myself reaching for something out of std.Io directly, the first
question that should probably be asked is "does a buffer make sense for
this?", and if the answer is no, then try to find a different way. :)