Glyph: fix advance/lsb fetches for higher glyph entries
This fixes things so that the advance for a glyph with an ID past the
numberOfHMetrics amount as given in maxp is gotten from the last
LongHorMetric entry. This is the correct method to do such things,
versus falling back to advanceWidthMax - I mis-interpreted the spec.

Specifically, this fixes rendering text with Annotation Mono, and
assumingly a number of other fonts exhibiting the same characteristics.

This additionally fixes reading of LSBs of said entries as the LSB index
was not being calculated relative to the numberOfHMetrics.