Polygon - introduce the working edge set
This introduces the concept of a working edge set to the rasterization
process, which aims to reduce the amount that the full set of
tessellated edges are iterated over for a particular scanline.

This will benefit drawing operations that have some combination of a
large amount of edges where each edge does not cover the entirety of the
y-axis of the draw area. Naturally, anti-aliased operations benefit more
due to the super-sampled co-ordinate space.

In brief, the model is as follows:

* A global working edge set stores both pointers to active edges and an
  equal amount of memory for calculated x-coordinates, and also
  pre-calculated directions (see below).
* Edge top and bottom co-ordinates are collected as a series of
  breakpoints.
* xEdgesForY now only runs on these breakpoints, updating the set of
  active edges.
* Edge x-coordinates are still calculated, sorted, and filtered each
  scanline. This section has been tightened a bit to just use
  symmetrical memory versus a MultiArrayList (allows us to manually
  slice and gives us better control over the memory).

Note that in an effort to better align the global edge set, some changes
have also been made to what an edge looks like. Direction has been
removed from edge, and that, along with top and bottom, are now
conditionally returned through helpers (y0 and y1 are stored now with no
regard to direction). This allows the full edge top/bottom/start/inc
pair to be stored in a proper power-of-two aligned fashion (32 bytes
instead of 40). Testing has actually shown a slight advantage to this
approach in terms of cycle count and possibly much better utilization of
the branch predictor (much lower branch miss count). We still store a
cached direction in the active edge set because, funny enough, removing
this and relying fully on the helper actually lowers performance again
(even though we get even better branch prediction performance, funny
enough).