painter: fast-path source-exclusive and clear operations
This sets up the painter for an optimization that, frankly, should have
been made a long time ago - fast-tracking opaque operations in the
event that they can be.

Now, when drawing with MSAA (the default currently) or no anti-aliasing
at all, when the source is fully opaque and/or the operation is
effectively a source-exclusive one, the compositor will fast-path to
just doing a memset on the surface via paintStride.

In MSAA, we also skip the vectored compositor for partial opacity in
these scenarios, and instead composite using a simple two-step scalar
operation; as the strides in these situations are likely are the ends
and do not really benefit from vectorization, given that they are
usually significantly smaller than the vector size itself (1-2 pixels
maybe).

Additionally, clear operations are now fast-pathed to just a memset of
transparent/opaque black (surface dependent).

This results in significant performance gains in both the MSAA and no
anti-aliasing scenarios for these trivial cases (painting with a single
solid color).

Not affected by this change: using gradient sources, or non-opaque
sources for all operators other than .src or .clear. These need to go
through the compositor pipeline still to do the appropriate operations.
SSAA is also not affected; while the mask is built using paintStride,
the nature of the operation in its entirety still demands everything go
through the compositor pipeline after the mask is built and downsampled.