painter: fast-path to memset for fully opaque operations
This sets up the painter for an optimization that, frankly, should have
been made a long time ago - fast-tracking opaque operations in the
event that they can be.

Now, when drawing with MSAA (the default currently) or no anti-aliasing
at all, when the source is fully opaque and the operation is effectively
a source-exclusive one, the compositor will fast-path to just doing a
memset on the surface via paintStride.

This results in significant performance gains in both the MSAA and no
anti-aliasing scenarios for these trivial cases (painting with a single
solid color).

Not affected by this change: using gradient sources or non-opaque
sources, as these need to go through the compositor pipeline still to do
the appropriate operations. SSAA is also not affected; while the mask is
built using paintStride, the nature of the operation in its entirety
still demands everything go through the compositor pipeline after the
mask is built and downsampled.

Note that this change is in its early stages and will be built upon -
more operators will be identified, and other scenarios where we can
fast-path the source for non-opaque sources will be identified as well
(e.g., .src can be for non-opaque because the source completely replaces
the destination).