sparse_coverage: make splitInner loop break conditions consistent
There was some inconsistencies here that were a relic of adapting the
code from tiny-skia. This is just a small cleanup to put the break
conditions into the same area (so now break condition is effectively <=
current_len) and ensure that both conditions break the loop (rather than
one return).