Polygon: do working edge set in-place
This removes the indirection that was done in maintaining a separate
slice of pointers to reference active edges in the working edge set.
Rather, we now just move the active edges to the front of the global set
and then slice off of that.

While the performance increase here in time is minimal, we are tracking
an increase in cache performance, in addition to the fact that we don't
have to store a slice of pointer-sized integers anymore as well.